+++ extensions/moralis-streams.env
@@ -1,0 +1,2 @@
+LOCATION=us-central1
+MORALIS_API_KEY=projects/284724398769/secrets/ext-moralis-streams-MORALIS_API_KEY/versions/1
