+++ package.json
@@ -25,6 +25,7 @@
     "react-dom": "^16.14.0",
     "react-intersection-observer": "^8.28.5",
     "react-moralis": "^1.4.2",
+    "react-router-dom": "^6.4.3",
     "react-scripts": "3.4.3",
     "react-spring": "^8.0.27",
     "react-three-fiber": "^4.2.21",
